projetos/process-ia-platform/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-multipart==0.0.12
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
reportlab==4.2.5
openai==1.78.1
alembic==1.13.3
email-validator==2.2.0
